scripts/auto-research/pyproject.toml
[project]
name = "auto-research"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "requests>=2.31",
    "textstat>=0.7",
]

[project.scripts]
auto-research = "research:main"
